Add Mounted SSL Support to NGINX Image
You can now mount an SSL certificate directory in the NGINX image similar to how you do it in the Apache image. By default, a CRT and KEY file is included in the image, but since it is generic, it might not play nice with some clients. It won't stop HTTP from working, but to get the best results with HTTPS, consider generating your own certificates and mounting them.